Module konst::cmp

Comparisong-related items.

Re-exports

pub use crate::polymorphism::kinds::IsNotStdKind;

pub use crate::polymorphism::kinds::IsRefKind;

pub use crate::polymorphism::kinds::IsStdKind;

Structs

CmpWrapper

A wrapper type for std types, which defines const_eq and const_cmp methods for them.

IsAConstCmp

Hack used to automatically wrap standard library types inside CmpWrapper, while leaving user defined types unwrapped.

Traits

ConstCmp

Marker trait for types that implement the const comparison methods.

ConstCmpUnref

A helper trait of ConstCmp, used for dereferencing.