Struct konst::cmp::CmpWrapper

pub struct CmpWrapper<T>(pub T);

A wrapper type for std types, which defines const_eq and const_cmp methods for them.

This is what coerce_to_cmp and the comparison macros convert standard library types into.

Example

use konst::{
    cmp::CmpWrapper,
    coerce_to_cmp,
};

use std::cmp::Ordering;

{
    // The `CmpWrapper<u32>` type annotation is just for the reader
    let foo: CmpWrapper<u32> = coerce_to_cmp!(10u32);
    assert!( foo.const_eq(&10));
    assert!(!foo.const_eq(&20));
     
    assert_eq!(foo.const_cmp(&5), Ordering::Greater);
    assert_eq!(foo.const_cmp(&10), Ordering::Equal);
    assert_eq!(foo.const_cmp(&15), Ordering::Less);
}
{
    let bar = CmpWrapper(Ordering::Equal);
    assert!( bar.const_eq(&Ordering::Equal));
    assert!(!bar.const_eq(&Ordering::Less));
     
    assert_eq!(bar.const_cmp(&Ordering::Less), Ordering::Greater);
    assert_eq!(bar.const_cmp(&Ordering::Equal), Ordering::Equal);
    assert_eq!(bar.const_cmp(&Ordering::Greater), Ordering::Less);
}

Tuple Fields
0: T
Implementations

impl<'a, T> CmpWrapper<&'a [T]>

pub const fn slice(x: &'a [T]) -> Self

For constructing from a reference to an array.

With slices you can do CmpWrapper(slice) as well.

impl CmpWrapper<i8>

pub const fn const_eq(self, other: &i8) -> bool

Compares self and other for equality.

impl CmpWrapper<u8>

pub const fn const_eq(self, other: &u8) -> bool

Compares self and other for equality.

impl CmpWrapper<i16>

pub const fn const_eq(self, other: &i16) -> bool

Compares self and other for equality.

impl CmpWrapper<u16>

pub const fn const_eq(self, other: &u16) -> bool

Compares self and other for equality.

impl CmpWrapper<i32>

pub const fn const_eq(self, other: &i32) -> bool

Compares self and other for equality.

impl CmpWrapper<u32>

pub const fn const_eq(self, other: &u32) -> bool

Compares self and other for equality.

impl CmpWrapper<i64>

pub const fn const_eq(self, other: &i64) -> bool

Compares self and other for equality.

impl CmpWrapper<u64>

pub const fn const_eq(self, other: &u64) -> bool

Compares self and other for equality.

impl CmpWrapper<i128>

pub const fn const_eq(self, other: &i128) -> bool

Compares self and other for equality.

impl CmpWrapper<u128>

pub const fn const_eq(self, other: &u128) -> bool

Compares self and other for equality.

impl CmpWrapper<isize>

pub const fn const_eq(self, other: &isize) -> bool

Compares self and other for equality.

impl CmpWrapper<usize>

pub const fn const_eq(self, other: &usize) -> bool

Compares self and other for equality.

impl CmpWrapper<bool>

pub const fn const_eq(self, other: &bool) -> bool

Compares self and other for equality.

impl CmpWrapper<char>

pub const fn const_eq(self, other: &char) -> bool

Compares self and other for equality.

impl CmpWrapper<u8>

pub const fn const_cmp(&self, other: &u8) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<u16>

pub const fn const_cmp(&self, other: &u16) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<u32>

pub const fn const_cmp(&self, other: &u32) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<u64>

pub const fn const_cmp(&self, other: &u64) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<u128>

pub const fn const_cmp(&self, other: &u128) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<usize>

pub const fn const_cmp(&self, other: &usize) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<i8>

pub const fn const_cmp(&self, other: &i8) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<i16>

pub const fn const_cmp(&self, other: &i16) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<i32>

pub const fn const_cmp(&self, other: &i32) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<i64>

pub const fn const_cmp(&self, other: &i64) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<i128>

pub const fn const_cmp(&self, other: &i128) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<isize>

pub const fn const_cmp(&self, other: &isize) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<bool>

pub const fn const_cmp(&self, other: &bool) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<char>

pub const fn const_cmp(&self, other: &char) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<u8>>

pub const fn const_eq(&self, other: &Option<u8>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<u8>>

pub const fn const_cmp(&self, other: &Option<u8>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<i8>>

pub const fn const_eq(&self, other: &Option<i8>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<i8>>

pub const fn const_cmp(&self, other: &Option<i8>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<u16>>

pub const fn const_eq(&self, other: &Option<u16>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<u16>>

pub const fn const_cmp(&self, other: &Option<u16>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<i16>>

pub const fn const_eq(&self, other: &Option<i16>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<i16>>

pub const fn const_cmp(&self, other: &Option<i16>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<u32>>

pub const fn const_eq(&self, other: &Option<u32>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<u32>>

pub const fn const_cmp(&self, other: &Option<u32>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<i32>>

pub const fn const_eq(&self, other: &Option<i32>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<i32>>

pub const fn const_cmp(&self, other: &Option<i32>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<u64>>

pub const fn const_eq(&self, other: &Option<u64>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<u64>>

pub const fn const_cmp(&self, other: &Option<u64>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<i64>>

pub const fn const_eq(&self, other: &Option<i64>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<i64>>

pub const fn const_cmp(&self, other: &Option<i64>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<u128>>

pub const fn const_eq(&self, other: &Option<u128>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<u128>>

pub const fn const_cmp(&self, other: &Option<u128>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<i128>>

pub const fn const_eq(&self, other: &Option<i128>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<i128>>

pub const fn const_cmp(&self, other: &Option<i128>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<usize>>

pub const fn const_eq(&self, other: &Option<usize>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<usize>>

pub const fn const_cmp(&self, other: &Option<usize>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<isize>>

pub const fn const_eq(&self, other: &Option<isize>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<isize>>

pub const fn const_cmp(&self, other: &Option<isize>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<bool>>

pub const fn const_eq(&self, other: &Option<bool>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<bool>>

pub const fn const_cmp(&self, other: &Option<bool>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<char>>

pub const fn const_eq(&self, other: &Option<char>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<char>>

pub const fn const_cmp(&self, other: &Option<char>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Range<u8>>

pub const fn const_eq(&self, other: &Range<u8>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<u8>>

pub const fn const_eq(&self, other: &'a Range<u8>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<u16>>

pub const fn const_eq(&self, other: &Range<u16>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<u16>>

pub const fn const_eq(&self, other: &'a Range<u16>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<u32>>

pub const fn const_eq(&self, other: &Range<u32>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<u32>>

pub const fn const_eq(&self, other: &'a Range<u32>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<u64>>

pub const fn const_eq(&self, other: &Range<u64>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<u64>>

pub const fn const_eq(&self, other: &'a Range<u64>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<u128>>

pub const fn const_eq(&self, other: &Range<u128>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<u128>>

pub const fn const_eq(&self, other: &'a Range<u128>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<usize>>

pub const fn const_eq(&self, other: &Range<usize>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<usize>>

pub const fn const_eq(&self, other: &'a Range<usize>) -> bool

Compares self and other for equality.

impl CmpWrapper<Range<char>>

pub const fn const_eq(&self, other: &Range<char>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&'a Range<char>>

pub const fn const_eq(&self, other: &'a Range<char>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<u8>>

pub const fn const_eq(&self, other: &RangeInclusive<u8>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<u16>>

pub const fn const_eq(&self, other: &RangeInclusive<u16>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<u32>>

pub const fn const_eq(&self, other: &RangeInclusive<u32>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<u64>>

pub const fn const_eq(&self, other: &RangeInclusive<u64>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<u128>>

pub const fn const_eq(&self, other: &RangeInclusive<u128>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<usize>>

pub const fn const_eq(&self, other: &RangeInclusive<usize>) -> bool

Compares self and other for equality.

impl CmpWrapper<RangeInclusive<char>>

pub const fn const_eq(&self, other: &RangeInclusive<char>) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU8>

pub const fn const_eq(&self, other: &NonZeroU8) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU8>

pub const fn const_cmp(&self, other: &NonZeroU8) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroI8>

pub const fn const_eq(&self, other: &NonZeroI8) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroI8>

pub const fn const_cmp(&self, other: &NonZeroI8) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroU16>

pub const fn const_eq(&self, other: &NonZeroU16) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU16>

pub const fn const_cmp(&self, other: &NonZeroU16) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroI16>

pub const fn const_eq(&self, other: &NonZeroI16) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroI16>

pub const fn const_cmp(&self, other: &NonZeroI16) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroU32>

pub const fn const_eq(&self, other: &NonZeroU32) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU32>

pub const fn const_cmp(&self, other: &NonZeroU32) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroI32>

pub const fn const_eq(&self, other: &NonZeroI32) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroI32>

pub const fn const_cmp(&self, other: &NonZeroI32) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroU64>

pub const fn const_eq(&self, other: &NonZeroU64) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU64>

pub const fn const_cmp(&self, other: &NonZeroU64) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroI64>

pub const fn const_eq(&self, other: &NonZeroI64) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroI64>

pub const fn const_cmp(&self, other: &NonZeroI64) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroU128>

pub const fn const_eq(&self, other: &NonZeroU128) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroU128>

pub const fn const_cmp(&self, other: &NonZeroU128) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroI128>

pub const fn const_eq(&self, other: &NonZeroI128) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroI128>

pub const fn const_cmp(&self, other: &NonZeroI128) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroUsize>

pub const fn const_eq(&self, other: &NonZeroUsize) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroUsize>

pub const fn const_cmp(&self, other: &NonZeroUsize) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<NonZeroIsize>

pub const fn const_eq(&self, other: &NonZeroIsize) -> bool

Compares self and other for equality.

impl CmpWrapper<NonZeroIsize>

pub const fn const_cmp(&self, other: &NonZeroIsize) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroU8>>

pub const fn const_eq(&self, other: &Option<NonZeroU8>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroU8>>

pub const fn const_cmp(&self, other: &Option<NonZeroU8>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroI8>>

pub const fn const_eq(&self, other: &Option<NonZeroI8>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroI8>>

pub const fn const_cmp(&self, other: &Option<NonZeroI8>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroU16>>

pub const fn const_eq(&self, other: &Option<NonZeroU16>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroU16>>

pub const fn const_cmp(&self, other: &Option<NonZeroU16>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroI16>>

pub const fn const_eq(&self, other: &Option<NonZeroI16>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroI16>>

pub const fn const_cmp(&self, other: &Option<NonZeroI16>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroU32>>

pub const fn const_eq(&self, other: &Option<NonZeroU32>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroU32>>

pub const fn const_cmp(&self, other: &Option<NonZeroU32>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroI32>>

pub const fn const_eq(&self, other: &Option<NonZeroI32>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroI32>>

pub const fn const_cmp(&self, other: &Option<NonZeroI32>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroU64>>

pub const fn const_eq(&self, other: &Option<NonZeroU64>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroU64>>

pub const fn const_cmp(&self, other: &Option<NonZeroU64>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroI64>>

pub const fn const_eq(&self, other: &Option<NonZeroI64>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroI64>>

pub const fn const_cmp(&self, other: &Option<NonZeroI64>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroU128>>

pub const fn const_eq(&self, other: &Option<NonZeroU128>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroU128>>

pub const fn const_cmp(&self, other: &Option<NonZeroU128>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroI128>>

pub const fn const_eq(&self, other: &Option<NonZeroI128>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroI128>>

pub const fn const_cmp(&self, other: &Option<NonZeroI128>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroUsize>>

pub const fn const_eq(&self, other: &Option<NonZeroUsize>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroUsize>>

pub const fn const_cmp(&self, other: &Option<NonZeroUsize>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<NonZeroIsize>>

pub const fn const_eq(&self, other: &Option<NonZeroIsize>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<NonZeroIsize>>

pub const fn const_cmp(&self, other: &Option<NonZeroIsize>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Ordering>

pub const fn const_eq(&self, other: &Ordering) -> bool

Compares self and other for equality.

impl CmpWrapper<Ordering>

pub const fn const_cmp(&self, other: &Ordering) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<Option<Ordering>>

pub const fn const_eq(&self, other: &Option<Ordering>) -> bool

Compares self and other for equality.

impl CmpWrapper<Option<Ordering>>

pub const fn const_cmp(&self, other: &Option<Ordering>) -> Ordering

Compares self and other for ordering.

impl<T> CmpWrapper<PhantomData<T>>

pub const fn const_eq(&self, other: &PhantomData<T>) -> bool

Compares self and other for equality.

impl<T> CmpWrapper<PhantomData<T>>

pub const fn const_cmp(&self, other: &PhantomData<T>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<PhantomPinned>

pub const fn const_eq(&self, other: &PhantomPinned) -> bool

Compares self and other for equality.

impl CmpWrapper<PhantomPinned>

pub const fn const_cmp(&self, other: &PhantomPinned) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[u16]>

pub const fn const_eq(&self, other: &[u16]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[u16]>

pub const fn const_cmp(&self, other: &[u16]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[u32]>

pub const fn const_eq(&self, other: &[u32]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[u32]>

pub const fn const_cmp(&self, other: &[u32]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[u64]>

pub const fn const_eq(&self, other: &[u64]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[u64]>

pub const fn const_cmp(&self, other: &[u64]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[u128]>

pub const fn const_eq(&self, other: &[u128]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[u128]>

pub const fn const_cmp(&self, other: &[u128]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[usize]>

pub const fn const_eq(&self, other: &[usize]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[usize]>

pub const fn const_cmp(&self, other: &[usize]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[i8]>

pub const fn const_eq(&self, other: &[i8]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[i8]>

pub const fn const_cmp(&self, other: &[i8]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[i16]>

pub const fn const_eq(&self, other: &[i16]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[i16]>

pub const fn const_cmp(&self, other: &[i16]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[i32]>

pub const fn const_eq(&self, other: &[i32]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[i32]>

pub const fn const_cmp(&self, other: &[i32]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[i64]>

pub const fn const_eq(&self, other: &[i64]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[i64]>

pub const fn const_cmp(&self, other: &[i64]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[i128]>

pub const fn const_eq(&self, other: &[i128]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[i128]>

pub const fn const_cmp(&self, other: &[i128]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[isize]>

pub const fn const_eq(&self, other: &[isize]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[isize]>

pub const fn const_cmp(&self, other: &[isize]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[bool]>

pub const fn const_eq(&self, other: &[bool]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[bool]>

pub const fn const_cmp(&self, other: &[bool]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[char]>

pub const fn const_eq(&self, other: &[char]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[char]>

pub const fn const_cmp(&self, other: &[char]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[&str]>

pub const fn const_eq(&self, other: &[&str]) -> bool

Compares self and other for equality.

impl CmpWrapper<&[&str]>

pub const fn const_cmp(&self, other: &[&str]) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[&[u8]]>

pub const fn const_eq(&self, other: &[&[u8]]) -> bool

Compares self and other for equality.

impl CmpWrapper<&[&[u8]]>

pub const fn const_cmp(&self, other: &[&[u8]]) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [u16]>>

pub const fn const_eq(&self, other: &Option<&'a [u16]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [u16]>>

pub const fn const_cmp(&self, other: &Option<&'a [u16]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [u32]>>

pub const fn const_eq(&self, other: &Option<&'a [u32]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [u32]>>

pub const fn const_cmp(&self, other: &Option<&'a [u32]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [u64]>>

pub const fn const_eq(&self, other: &Option<&'a [u64]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [u64]>>

pub const fn const_cmp(&self, other: &Option<&'a [u64]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [u128]>>

pub const fn const_eq(&self, other: &Option<&'a [u128]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [u128]>>

pub const fn const_cmp(&self, other: &Option<&'a [u128]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [usize]>>

pub const fn const_eq(&self, other: &Option<&'a [usize]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [usize]>>

pub const fn const_cmp(&self, other: &Option<&'a [usize]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [i8]>>

pub const fn const_eq(&self, other: &Option<&'a [i8]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [i8]>>

pub const fn const_cmp(&self, other: &Option<&'a [i8]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [i16]>>

pub const fn const_eq(&self, other: &Option<&'a [i16]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [i16]>>

pub const fn const_cmp(&self, other: &Option<&'a [i16]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [i32]>>

pub const fn const_eq(&self, other: &Option<&'a [i32]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [i32]>>

pub const fn const_cmp(&self, other: &Option<&'a [i32]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [i64]>>

pub const fn const_eq(&self, other: &Option<&'a [i64]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [i64]>>

pub const fn const_cmp(&self, other: &Option<&'a [i64]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [i128]>>

pub const fn const_eq(&self, other: &Option<&'a [i128]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [i128]>>

pub const fn const_cmp(&self, other: &Option<&'a [i128]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [isize]>>

pub const fn const_eq(&self, other: &Option<&'a [isize]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [isize]>>

pub const fn const_cmp(&self, other: &Option<&'a [isize]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [bool]>>

pub const fn const_eq(&self, other: &Option<&'a [bool]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [bool]>>

pub const fn const_cmp(&self, other: &Option<&'a [bool]>) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [char]>>

pub const fn const_eq(&self, other: &Option<&'a [char]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [char]>>

pub const fn const_cmp(&self, other: &Option<&'a [char]>) -> Ordering

Compares self and other for ordering.

impl<'a, 'b> CmpWrapper<Option<&'a [&'b str]>>

pub const fn const_eq(&self, other: &Option<&'a [&'b str]>) -> bool

Compares self and other for equality.

impl<'a, 'b> CmpWrapper<Option<&'a [&'b str]>>

pub const fn const_cmp(&self, other: &Option<&'a [&'b str]>) -> Ordering

Compares self and other for ordering.

impl<'a, 'b> CmpWrapper<Option<&'a [&'b [u8]]>>

pub const fn const_eq(&self, other: &Option<&'a [&'b [u8]]>) -> bool

Compares self and other for equality.

impl<'a, 'b> CmpWrapper<Option<&'a [&'b [u8]]>>

pub const fn const_cmp(&self, other: &Option<&'a [&'b [u8]]>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&[u8]>

pub const fn const_eq(&self, other: &[u8]) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<&[u8]>

pub const fn const_cmp(&self, other: &[u8]) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a [u8]>>

pub const fn const_eq(&self, other: &Option<&'a [u8]>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a [u8]>>

pub const fn const_cmp(&self, other: &Option<&'a [u8]>) -> Ordering

Compares self and other for ordering.

impl CmpWrapper<&str>

pub const fn const_eq(&self, other: &str) -> bool

Compares self and other for equality.

impl CmpWrapper<&str>

pub const fn const_cmp(&self, other: &str) -> Ordering

Compares self and other for ordering.

impl<'a> CmpWrapper<Option<&'a str>>

pub const fn const_eq(&self, other: &Option<&'a str>) -> bool

Compares self and other for equality.

impl<'a> CmpWrapper<Option<&'a str>>

pub const fn const_cmp(&self, other: &Option<&'a str>) -> Ordering

Compares self and other for ordering.