konst

Module iter

Const equivalent of iterators with a specific next function signature.

The docs for IntoIterKind has more information on const equivalents of IntoIterator and Iterator.

Modules

iterator_dsl

Documentation on the iterator DSL that some konst::iter macros support.

Macros

collect_const

Collects an iterator constant into an array

eval

Emulates iterator method chains, by expanding to equivalent code.

for_each

Iterates over all elements of an iterator, const equivalent of [Iterator::for_each]

into_iter

Macro for converting IntoIterKind implementors into const iterators.

Structs

IntoIterWrapper

Wrapper for IntoIterKind implementors, that defines different methods depending on the value of <T as IntoIterKind>::Kind.

IsIntoIterKind

Marker type for proving that T: IntoIterKind<Kind = K>

IsIteratorKind

For marking some type as being an iterator in its IntoIterKind::Kind associated type.

IsNonIteratorKind

For marking some type as being convertible to an iterator in its IntoIterKind::Kind associated type.

IsStdKind

For marking some type as being from std in its IntoIterKind::Kind associated type.

Traits

IntoIterKind

Const analog of the [IntoIterator] trait.