Module konst::nonzero::cmp

const fns for comparing NonZero* types for equality and ordering.

Functions

cmp_nonzeroi8

Compares two NonZeroI8, returning the ordering of left relative to right.

cmp_nonzeroi16

Compares two NonZeroI16, returning the ordering of left relative to right.

cmp_nonzeroi32

Compares two NonZeroI32, returning the ordering of left relative to right.

cmp_nonzeroi64

Compares two NonZeroI64, returning the ordering of left relative to right.

cmp_nonzeroi128

Compares two NonZeroI128, returning the ordering of left relative to right.

cmp_nonzeroisize

Compares two NonZeroIsize, returning the ordering of left relative to right.

cmp_nonzerou8

Compares two NonZeroU8, returning the ordering of left relative to right.

cmp_nonzerou16

Compares two NonZeroU16, returning the ordering of left relative to right.

cmp_nonzerou32

Compares two NonZeroU32, returning the ordering of left relative to right.

cmp_nonzerou64

Compares two NonZeroU64, returning the ordering of left relative to right.

cmp_nonzerou128

Compares two NonZeroU128, returning the ordering of left relative to right.

cmp_nonzerousize

Compares two NonZeroUsize, returning the ordering of left relative to right.

cmp_option_nonzeroi8

Compares two Option<NonZeroI8>, returning the ordering of left relative to right.

cmp_option_nonzeroi16

Compares two Option<NonZeroI16>, returning the ordering of left relative to right.

cmp_option_nonzeroi32

Compares two Option<NonZeroI32>, returning the ordering of left relative to right.

cmp_option_nonzeroi64

Compares two Option<NonZeroI64>, returning the ordering of left relative to right.

cmp_option_nonzeroi128

Compares two Option<NonZeroI128>, returning the ordering of left relative to right.

cmp_option_nonzeroisize

Compares two Option<NonZeroIsize>, returning the ordering of left relative to right.

cmp_option_nonzerou8

Compares two Option<NonZeroU8>, returning the ordering of left relative to right.

cmp_option_nonzerou16

Compares two Option<NonZeroU16>, returning the ordering of left relative to right.

cmp_option_nonzerou32

Compares two Option<NonZeroU32>, returning the ordering of left relative to right.

cmp_option_nonzerou64

Compares two Option<NonZeroU64>, returning the ordering of left relative to right.

cmp_option_nonzerou128

Compares two Option<NonZeroU128>, returning the ordering of left relative to right.

cmp_option_nonzerousize

Compares two Option<NonZeroUsize>, returning the ordering of left relative to right.

eq_nonzeroi8

Compares two NonZeroI8 for equality.

eq_nonzeroi16

Compares two NonZeroI16 for equality.

eq_nonzeroi32

Compares two NonZeroI32 for equality.

eq_nonzeroi64

Compares two NonZeroI64 for equality.

eq_nonzeroi128

Compares two NonZeroI128 for equality.

eq_nonzeroisize

Compares two NonZeroIsize for equality.

eq_nonzerou8

Compares two NonZeroU8 for equality.

eq_nonzerou16

Compares two NonZeroU16 for equality.

eq_nonzerou32

Compares two NonZeroU32 for equality.

eq_nonzerou64

Compares two NonZeroU64 for equality.

eq_nonzerou128

Compares two NonZeroU128 for equality.

eq_nonzerousize

Compares two NonZeroUsize for equality.

eq_option_nonzeroi8

Compares two Option<NonZeroI8> for equality.

eq_option_nonzeroi16

Compares two Option<NonZeroI16> for equality.

eq_option_nonzeroi32

Compares two Option<NonZeroI32> for equality.

eq_option_nonzeroi64

Compares two Option<NonZeroI64> for equality.

eq_option_nonzeroi128

Compares two Option<NonZeroI128> for equality.

eq_option_nonzeroisize

Compares two Option<NonZeroIsize> for equality.

eq_option_nonzerou8

Compares two Option<NonZeroU8> for equality.

eq_option_nonzerou16

Compares two Option<NonZeroU16> for equality.

eq_option_nonzerou32

Compares two Option<NonZeroU32> for equality.

eq_option_nonzerou64

Compares two Option<NonZeroU64> for equality.

eq_option_nonzerou128

Compares two Option<NonZeroU128> for equality.

eq_option_nonzerousize

Compares two Option<NonZeroUsize> for equality.