Module konst::polymorphism

Miscelaneous items used for emulating polymorphism without trait methods.

typewit

This crate uses typewit for emulating trait based polymorphism in const fns, because trait methods cannot be called in const fns on stable(as of Rust 1.69.0)

Re-exports

pub use typewit;

pub use typewit::simple_type_witness;

pub use typewit::type_fn;

pub use typewit::type_fn;

pub use typewit::CallFn;

pub use typewit::HasTypeWitness;

pub use typewit::MakeTypeWitness;

pub use typewit::TypeEq;

pub use typewit::TypeFn;

pub use typewit::TypeWitnessTypeArg;

Modules

kinds

Markers used to classify types, used as Kind associated types.

Macros

type_eq_projection_fn

Declares a function for converting a TypeEq<L, R> to TypeEq<Foo<L>, Foo<R>>.