konst
Module ptr

Const equivalents of raw pointer and NonNull methods.

Removed in 0.3.0

These functions were removed in 0.3.0 because there is an equivalent const fn in the standard library:

deref: raw pointers can be dereferenced since Rust 1.58.0

deref_mut: the unstable feature for using mutable references also allows dereferencing mutable pointers.

Modules

nonnull

Const equivalents of NonNull methods.

Functions

as_ref⚠

Const equivalent of <*const>::as_ref

is_null

Const equivalent of <*const T>::is_null