konst::slice

Function slice_range

pub const fn slice_range<T>(slice: &[T], start: usize, end: usize) -> &[T]

A const equivalent of &slice[start..end].

If start >= end or slice.len() < start , this returns an empty slice.

If slice.len() < end, this returns the slice from start.

Alternatives

For a const equivalent of &slice[start..] there’s slice_from.

For a const equivalent of &slice[..end] there’s slice_up_to.

Performance

If the “rust_1_64” feature is disabled, thich takes linear time to remove the leading and trailing elements, proportional to start + (slice.len() - end).

If the “rust_1_64” feature is enabled, it takes constant time to run.

Example

use konst::slice::slice_range;

const FIBB: &[u16] = &[3, 5, 8, 13, 21, 34, 55, 89];

const TWO: &[u16] = slice_range(FIBB, 2, 4);
const FOUR: &[u16] = slice_range(FIBB, 4, 7);
const NONE: &[u16] = slice_range(FIBB, 0, 0);
const ALL: &[u16] = slice_range(FIBB, 0, 1000);

assert_eq!(TWO, &[8, 13]);
assert_eq!(FOUR, &[21, 34, 55]);
assert_eq!(NONE, &[]);
assert_eq!(ALL, FIBB);