Module konst::parsing

Parsing using const fn methods.

You can use the Parser type to parse from string, more information in its documentation.

If you’re looking for functions to parse some type from an entire string (instead of only part of it), then you want to look in the module for that type, eg: primitive::parse_bool.

If you do want to parse a type fron only part of a string, then you can use Parser’s parse_* methods, or the parse_with macro.

Structs

ParseError

Error returned by all parsing methods that return Result.

Parser

For parsing and traversing over strings in const contexts.

StdParser

Parses a standard library type, determined by the StdType type parameter.

Enums

ErrorKind

What kind of parsing error this is.

ParseDirection

The direction that a parser was parsing from when an error happened.

Traits

HasParser

Gets a type that parses Self with a parse_with method.

Type Definitions

ParseValueResult

Result alias for functions that parse values.

ParserResult

Result alias for functions that mutate the parser fallibly.