Module konst::primitive::cmp

const fns for comparing primitive types for equality and ordering.

Functions

cmp_bool

Compares two bool, returning the ordering of left relative to right.

cmp_char

Compares two char, returning the ordering of left relative to right.

cmp_i8

Compares two i8, returning the ordering of left relative to right.

cmp_i16

Compares two i16, returning the ordering of left relative to right.

cmp_i32

Compares two i32, returning the ordering of left relative to right.

cmp_i64

Compares two i64, returning the ordering of left relative to right.

cmp_i128

Compares two i128, returning the ordering of left relative to right.

cmp_isize

Compares two isize, returning the ordering of left relative to right.

cmp_option_bool

Compares two Option<bool>, returning the ordering of left relative to right.

cmp_option_char

Compares two Option<char>, returning the ordering of left relative to right.

cmp_option_i8

Compares two Option<i8>, returning the ordering of left relative to right.

cmp_option_i16

Compares two Option<i16>, returning the ordering of left relative to right.

cmp_option_i32

Compares two Option<i32>, returning the ordering of left relative to right.

cmp_option_i64

Compares two Option<i64>, returning the ordering of left relative to right.

cmp_option_i128

Compares two Option<i128>, returning the ordering of left relative to right.

cmp_option_isize

Compares two Option<isize>, returning the ordering of left relative to right.

cmp_option_u8

Compares two Option<u8>, returning the ordering of left relative to right.

cmp_option_u16

Compares two Option<u16>, returning the ordering of left relative to right.

cmp_option_u32

Compares two Option<u32>, returning the ordering of left relative to right.

cmp_option_u64

Compares two Option<u64>, returning the ordering of left relative to right.

cmp_option_u128

Compares two Option<u128>, returning the ordering of left relative to right.

cmp_option_usize

Compares two Option<usize>, returning the ordering of left relative to right.

cmp_u8

Compares two u8, returning the ordering of left relative to right.

cmp_u16

Compares two u16, returning the ordering of left relative to right.

cmp_u32

Compares two u32, returning the ordering of left relative to right.

cmp_u64

Compares two u64, returning the ordering of left relative to right.

cmp_u128

Compares two u128, returning the ordering of left relative to right.

cmp_usize

Compares two usize, returning the ordering of left relative to right.

eq_option_bool

Compares two Option<bool> for equality.

eq_option_char

Compares two Option<char> for equality.

eq_option_i8

Compares two Option<i8> for equality.

eq_option_i16

Compares two Option<i16> for equality.

eq_option_i32

Compares two Option<i32> for equality.

eq_option_i64

Compares two Option<i64> for equality.

eq_option_i128

Compares two Option<i128> for equality.

eq_option_isize

Compares two Option<isize> for equality.

eq_option_u8

Compares two Option<u8> for equality.

eq_option_u16

Compares two Option<u16> for equality.

eq_option_u32

Compares two Option<u32> for equality.

eq_option_u64

Compares two Option<u64> for equality.

eq_option_u128

Compares two Option<u128> for equality.

eq_option_usize

Compares two Option<usize> for equality.