Module konst::result

const equivalents of Result methods.

Macros

and_then

A const equivalent of Result::and_then

err

A const equivalent of Result::err

map

A const equivalent of Result::map

map_err

A const equivalent of Result::map_err

ok

A const equivalent of Result::ok

or_else

A const equivalent of Result::or_else

unwrap_ctx

For unwrapping Results in const contexts with some error message.

unwrap_err_or_else

Returns the error in the Err variant, otherwise runs a closure/function with the value in the Ok variant.

unwrap_or

A const equivalent of Result::unwrap_or

unwrap_or_else

A const equivalent of Result::unwrap_or_else