Trait konst::slice::BytesPattern

pub trait BytesPattern<const N: usize>: HasTypeWitness<BytesPatternInput<N, Self>> { }

A byte slice pattern.

Types that implement this trait can be used to search into a byte slice.

This trait can only be implemented in the konst crate.